backend/requirements.txt
# -----------------------------------------------------------------------------
# EAR-API (Embodied Aesthetic Reconstruction)
# Multi-Museum FastAPI backend
# -----------------------------------------------------------------------------
# ✅ Tested with Python 3.10–3.11
# ✅ Works on macOS (MPS), Windows, Linux (CUDA)
# ✅ Compatible with VS Code / Hugging Face / Render
# -----------------------------------------------------------------------------

# --- Core API Framework ---
fastapi==0.115.2
uvicorn==0.32.0

# --- Numerical & Data Handling ---
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0

# --- Computer Vision / Pose Estimation ---
ultralytics==8.3.34
opencv-python==4.10.0.84

# --- CLIP Embedding (OpenAI / LAION) ---
open_clip_torch==2.24.0
torch==2.3.1
torchvision==0.18.1

# --- Optional Utilities (recommended for extended scripts) ---
tqdm==4.66.4     # progress bars in build_embeddings.py
